Refactoring use commun code between option, result and accum
